Modify count,payload to u32 and reset written_records to 0 on MAX.
Modified count, payload to u32 as they were ( specifically count ) was
getting wrap around if the payload length in the frame is of a greater
value ( > 60 K) . This resulted in the count getting recalculated with
a different offset , due to wrap around .This scenario resulted in to
the while loop never to exit due to count getting improperly calculated.
Avoid this by changing the respective variables to u32

CRs-Fixed: 2708208

+++ qcwcn/wifi_hal/wifilogger_diag.cpp
@@ -853,10 +853,10 @@ static wifi_error process_beacon_received_event(hal_info *info,
     return status;
 }
 
-static wifi_error process_fw_diag_msg(hal_info *info, u8* buf, u16 length)
+static wifi_error process_fw_diag_msg(hal_info *info, u8* buf, u32 length)
 {
-    u16 count = 0, id;
-    u16 payloadlen = 0;
+    u32 count = 0, id;
+    u32 payloadlen = 0;
     u16 hdr_size = 0;
     wifi_error status;
     fw_diag_msg_fixed_hdr_t *diag_msg_fixed_hdr;
@@ -2815,7 +2815,7 @@ wifi_error diag_message_handler(hal_info *info, nl_msg *msg)
         diag_fw_type = event_hdr->diag_type;
         if (diag_fw_type == DIAG_TYPE_FW_MSG) {
             dbglog_slot *slot;
-            u16 length = 0;
+            u32 length = 0;
 
             slot = (dbglog_slot *)buf;
             length = get_le32((u8 *)&slot->length);
